Move mirror types to core
