--- agent/src/main/scala/io/scalac/mesmer/agent/akka/mirror/AkkaMirror.scala
@@ -1,6 +1,0 @@
-package akka
-
-object AkkaMirror {
-  type ActorRefWithCell = akka.actor.ActorRefWithCell
-  type Cell             = akka.actor.Cell
-}
