--- agent/src/main/scala/io/scalac/mesmer/agent/akka/stream/impl/AkkaMirrorTypes.scala
@@ -1,17 +1,0 @@
-package akka
-
-import akka.stream.impl.ExtendedActorMaterializer
-import akka.stream.impl.fusing.GraphInterpreter
-import akka.stream.impl.fusing.GraphInterpreterShell
-import akka.stream.stage.GraphStageLogic
-import akka.util.{ OptionVal => AkkaOptionVal }
-
-object AkkaMirrorTypes {
-
-  type ExtendedActorMaterializerMirror = ExtendedActorMaterializer
-  type GraphInterpreterMirror          = GraphInterpreter
-  type GraphInterpreterShellMirror     = GraphInterpreterShell
-  type GraphStageLogicMirror           = GraphStageLogic
-  type OptionVal[+T >: Null]           = AkkaOptionVal[T]
-
-}
