+++ agent/src/main/scala/io/scalac/mesmer/agent/akka/stream/impl/PhasedFusingActorMaterializerAdvice.scala
@@ -1,6 +1,5 @@
 package io.scalac.mesmer.agent.akka.stream.impl
 
-import akka.AkkaMirrorTypes
 import akka.actor.ActorRef
 import akka.actor.typed.scaladsl.adapter._
 import net.bytebuddy.asm.Advice._
@@ -14,6 +13,6 @@ object PhasedFusingActorMaterializerAdvice {
 
   @OnMethodExit
   def actorOf(@Return ref: ActorRef, @This self: Object): Unit =
-    EventBus(self.asInstanceOf[AkkaMirrorTypes.ExtendedActorMaterializerMirror].system.toTyped)
+    EventBus(self.asInstanceOf[akka.MesmerMirrorTypes.ExtendedActorMaterializerMirror].system.toTyped)
       .publishEvent(ActorEvent.TagsSet(ActorRefTags(ref, Set(Tag.stream))))
 }
